pyproject.toml
[project]
name = "smartclm-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.41.0",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "ragas>=0.2.15",
    "seaborn>=0.13.2",
    "streamlit>=1.46.1",
    "httpx>=0.27.0",
    "pandas>=2.0.0",
    "fastapi>=0.100.0",
    "aiofiles>=23.0.0",
    "python-multipart>=0.0.6",
    "sqlmodel>=0.0.24",
    "asyncpg>=0.30.0",
    "pgvector>=0.4.1",
    "langchain>=0.3.26",
    "inflection>=0.5.1",
    "langchain-aws>=0.2.0,<0.2.28",
    "aioboto3>=13.0.0",
    "python-dotenv>=1.0.0",
]
